requirements.txt
fastapi
uvicorn
langchain
langchain-openai
openai
python-dotenv
langchain-community
beautifulsoup4
faiss-cpu
